// MEMAC/src/types/Representative.ts
export interface Representative {
    id: string;
    fullName: string;
    shortName: string;
    phone: string;
    city: string;
    state: string;
    document: string;
    contact: string;
    address?: string;
    neighborhood?: string;
    zipCode?: string;
    email?: string;
    commission?: string;
    notes?: string;
  }
  